Use size_t type for loop indices where appropriate.

+++ library/src/enummixed/enummixed.cc
@@ -102,9 +102,9 @@ EnumMixedStrategySolver<T>::SolveDetailed(const Game &p_game) const
   Matrix<T> A2(1, p_game->Players()[2]->Strategies().size(),
 	       1, p_game->Players()[1]->Strategies().size());
 
-  for (int i = 1; i <= p_game->Players()[1]->Strategies().size(); i++) {
+  for (size_t i = 1; i <= p_game->Players()[1]->Strategies().size(); i++) {
     profile->SetStrategy(p_game->Players()[1]->Strategies()[i]);
-    for (int j = 1; j <= p_game->Players()[2]->Strategies().size(); j++) {
+    for (size_t j = 1; j <= p_game->Players()[2]->Strategies().size(); j++) {
       profile->SetStrategy(p_game->Players()[2]->Strategies()[j]);
       A1(i, j) = fac * (profile->GetPayoff(1) - min);
       A2(j, i) = fac * (profile->GetPayoff(2) - min);
@@ -143,13 +143,13 @@ EnumMixedStrategySolver<T>::SolveDetailed(const Game &p_game) const
       // check if solution is nash 
       // need only check complementarity, since it is feasible
       bool nash = true;
-      for (int k = 1; nash && k <= p_game->Players()[1]->Strategies().size(); k++) {
+      for (size_t k = 1; nash && k <= p_game->Players()[1]->Strategies().size(); k++) {
 	if (bfs1.count(k) && bfs2.count(-k)) {
 	  nash = nash && EqZero(bfs1[k] * bfs2[-k]);
 	}
       }
 
-      for (int k = 1; nash && k <= p_game->Players()[2]->Strategies().size(); k++) {
+      for (size_t k = 1; nash && k <= p_game->Players()[2]->Strategies().size(); k++) {
 	if (bfs2.count(k) && bfs1.count(-k)) {
 	  nash = nash && EqZero(bfs2[k] * bfs1[-k]);
 	}
@@ -158,12 +158,12 @@ EnumMixedStrategySolver<T>::SolveDetailed(const Game &p_game) const
       if (nash) {
 	MixedStrategyProfile<T> profile(p_game->NewMixedStrategyProfile(static_cast<T>(0)));
 	static_cast<Vector<T> &>(profile) = static_cast<T>(0);
-	for (int k = 1; k <= p_game->Players()[1]->Strategies().size(); k++) {
+	for (size_t k = 1; k <= p_game->Players()[1]->Strategies().size(); k++) {
 	  if (bfs1.count(k)) {
 	    profile[p_game->Players()[1]->Strategies()[k]] = -bfs1[k];
 	  }
 	} 
-	for (int k = 1; k <= p_game->Players()[2]->Strategies().size(); k++) {
+	for (size_t k = 1; k <= p_game->Players()[2]->Strategies().size(); k++) {
 	  if (bfs2.count(k)) {
 	    profile[p_game->Players()[2]->Strategies()[k]] = -bfs2[k];
 	  }
